requirements.txt
grpcio~=1.71.0
requests~=2.32.3